+++ packages/atproto/test/src/servers/servers_service_test.dart
@@ -17,6 +17,7 @@ import 'package:atproto/src/entities/current_session.dart';
 import 'package:atproto/src/entities/email_update.dart';
 import 'package:atproto/src/entities/invite_codes.dart';
 import 'package:atproto/src/entities/server_info.dart';
+import 'package:atproto/src/entities/signing_key.dart';
 import 'package:atproto/src/servers/servers_service.dart';
 
 void main() {
@@ -1304,6 +1305,68 @@ void main() {
           email: 'contact@shinyakato.dev',
           token: 'xxxxxx',
         ),
+      );
+    });
+  });
+
+  group('.createSigningKey', () {
+    test('normal case', () async {
+      final servers = ServersService(
+        did: 'test',
+        protocol: core.Protocol.https,
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/servers/data/create_signing_key.json',
+        ),
+      );
+
+      final response = await servers.createSigningKey();
+
+      expect(response, isA<core.XRPCResponse>());
+      expect(response.data, isA<SigningKey>());
+    });
+
+    test('when unauthorized', () async {
+      final servers = ServersService(
+        did: 'test',
+        protocol: core.Protocol.https,
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/data/error.json',
+          statusCode: 401,
+        ),
+      );
+
+      atp_test.expectUnauthorizedException(
+        () async => await servers.createSigningKey(),
+      );
+    });
+
+    test('when rate limit exceeded', () async {
+      final servers = ServersService(
+        did: 'test',
+        protocol: core.Protocol.https,
+        service: 'test',
+        context: core.ClientContext(
+          accessJwt: '1234',
+          timeout: Duration.zero,
+        ),
+        mockedPostClient: atp_test.createMockedPostClient(
+          'test/src/data/error.json',
+          statusCode: 429,
+        ),
+      );
+
+      atp_test.expectRateLimitExceededException(
+        () async => await servers.createSigningKey(),
       );
     });
   });